refactor(worker): U6 — split repository.js into focused modules
Extract 4 modules from the 9163-line monolith:
- membership-repository.js (211 lines): role gates, membership queries
- bootstrap-repository.js (267 lines): BOOTSTRAP_CAPACITY_VERSION,
  revision hash, capacity meta, selected-learner resolver
- repository-helpers.js (76 lines): shared pure utilities
- mutation-repository.js (401 lines): CAS mutation, receipts,
  withLearnerMutation/withAccountMutation orchestrators

Pure code movement — zero behaviour change. repository.js retains
barrel re-exports so all existing consumers work without import path
changes. BOOTSTRAP_CAPACITY_VERSION bump rule restated in header.
No circular dependencies. 4065 tests pass, audit:client green.

Closes R6.
