+++ worker/src/repository.js
@@ -5,7 +5,6 @@ import {
   filterSessions,
   gameStateKey,
   normaliseEventLog,
-  normaliseLearnerRecord,
   normaliseLearnersSnapshot,
   normalisePracticeSessionRecord,
   normaliseRepositoryBundle,
@@ -26,12 +25,71 @@ import {
   resolvePostMegaSeedShape,
 } from '../../shared/spelling/post-mastery-seed-shapes.js';
 import {
-  PLATFORM_ROLES,
-  canManageAccountRoles,
-  canViewAdminHub,
-  canViewParentHub,
   normalisePlatformRole,
 } from '../../src/platform/access/roles.js';
+// P3 U6: membership, role-gate, and access-control helpers extracted to
+// a focused module. Barrel re-exports at the bottom of this file keep
+// all existing consumers working without import-path changes.
+import {
+  accountPlatformRole,
+  accountType,
+  getMembership,
+  learnerRowToRecord,
+  listMembershipRows,
+  MEMBERSHIP_ROLES,
+  membershipRowToModel,
+  normaliseRequestedPlatformRole,
+  requireAccountRoleManager,
+  requireAdminHubAccess,
+  requireGrammarTransferAdmin,
+  requireLearnerReadAccess,
+  requireLearnerWriteAccess,
+  requireMonsterVisualConfigManager,
+  requireParentHubAccess,
+  requireSubjectContentExportAccess,
+  requireSubjectContentWriteAccess,
+  writableRole,
+} from './membership-repository.js';
+// P3 U6: bootstrap constants, revision hash, capacity meta, and
+// selected-learner resolver extracted to a focused module.
+import {
+  bootstrapCapacityMeta,
+  BOOTSTRAP_CAPACITY_VERSION,
+  BOOTSTRAP_MODES,
+  BOOTSTRAP_V2_ENVELOPE_SHAPE,
+  compactLearnerListEntry,
+  computeBootstrapRevisionHash,
+  computeWritableLearnerStatesDigest,
+  PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LOOKUP_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_CAPACITY_VERSION,
+  PUBLIC_BOOTSTRAP_RECENT_EVENT_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_RECENT_SESSION_LIMIT_PER_LEARNER,
+  resolveBootstrapSelectedLearnerId,
+} from './bootstrap-repository.js';
+// P3 U6: shared pure-utility helpers extracted to a focused module.
+import {
+  asTs,
+  isMissingTableError,
+  isPlainObject,
+  logMutation,
+  MUTATION_POLICY_VERSION,
+  mutationPayloadHash,
+  safeJsonParse,
+  stableClone,
+  stableStringify,
+} from './repository-helpers.js';
+// P3 U6: mutation envelope, receipt persistence, and CAS orchestrators.
+import {
+  buildMutationMeta,
+  idempotencyReuseError,
+  loadMutationReceipt,
+  normaliseMutationInput,
+  staleWriteError,
+  storeMutationReceipt,
+  withAccountMutation,
+  withLearnerMutation,
+} from './mutation-repository.js';
 import { buildAdminHubReadModel } from '../../src/platform/hubs/admin-read-model.js';
 import { buildParentHubReadModel } from '../../src/platform/hubs/parent-read-model.js';
 import { monsterIdForSpellingWord } from '../../src/platform/game/monster-system.js';
@@ -98,9 +156,8 @@ import {
 } from './d1.js';
 import { getReadModelDerivedWriteBreaker } from './circuit-breaker-server.js';
 
-const WRITABLE_MEMBERSHIP_ROLES = new Set(['owner', 'member']);
-const MEMBERSHIP_ROLES = new Set(['owner', 'member', 'viewer']);
-const MUTATION_POLICY_VERSION = 1;
+// WRITABLE_MEMBERSHIP_ROLES / MEMBERSHIP_ROLES → membership-repository.js
+// MUTATION_POLICY_VERSION → repository-helpers.js
 const PUBLIC_SPELLING_YEAR_LABELS = new Map([
   ['3-4', 'Years 3-4'],
   ['5-6', 'Years 5-6'],
@@ -118,64 +175,8 @@ const PUBLIC_EVENT_TYPES = new Set([
   'reward.monster',
   'platform.practice-streak-hit',
 ]);
-const PUBLIC_BOOTSTRAP_RECENT_SESSION_LIMIT_PER_LEARNER = 5;
-const PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LIMIT_PER_LEARNER = 1;
-const PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LOOKUP_LIMIT_PER_LEARNER = 5;
-const PUBLIC_BOOTSTRAP_RECENT_EVENT_LIMIT_PER_LEARNER = 50;
-// U7: bumped from 1 → 2 when the selected-learner-bounded envelope landed.
-// U1 hotfix follow-up 2026-04-26: bumped 2 → 3 in the same PR that:
-//   - adds `bootstrapCapacity.subjectStatesBounded` (required-field addition
-//     per the capacity release-gate plan, docs/plans/2026-04-25-002...),
-//   - extends the revision-hash input set with
-//     `writableLearnerStatesDigest` (a state_revision digest across every
-//     writable learner) so sibling subject_state writes invalidate the
-//     `bootstrapNotModifiedProbe` short-circuit (B1 blocker).
-// Stale v2 clients will naturally miss, re-fetch, and bind to the v3 hash.
-// Any additive required field on the bootstrap envelope MUST bump this in
-// the same PR. `tests/worker-bootstrap-v2.test.js` has a snapshot test that
-// fails if the envelope shape changes without a version bump (scenario 15).
-const PUBLIC_BOOTSTRAP_CAPACITY_VERSION = 3;
-export const BOOTSTRAP_CAPACITY_VERSION = PUBLIC_BOOTSTRAP_CAPACITY_VERSION;
-
-// U7: closed union for `meta.capacity.bootstrapMode` when the public
-// bootstrap runs. `full-legacy` covers the `publicReadModels=false` path
-// (non-public internal callers still go through the unrestricted bundle).
-// `not-modified` is returned when the client's `lastKnownRevision` matches
-// the current server hash and we return a < 2 KB short response.
-export const BOOTSTRAP_MODES = new Set([
-  'selected-learner-bounded',
-  'full-legacy',
-  'not-modified',
-]);
-
-// U7: snapshot for the v2 envelope shape. Locked per-version; a required
-// shape change without a `BOOTSTRAP_CAPACITY_VERSION` bump + a snapshot
-// update in the same PR fails the release-rule test (scenario 15).
-// EVIDENCE_SCHEMA_VERSION is deliberately NOT bumped — that constant
-// covers the capacity evidence doc schema (U3), not the bootstrap
-// envelope; bootstrap envelope evolution is governed by its own version.
-export const BOOTSTRAP_V2_ENVELOPE_SHAPE = Object.freeze({
-  version: PUBLIC_BOOTSTRAP_CAPACITY_VERSION,
-  requiredTopLevelKeys: Object.freeze([
-    'account',
-    'eventLog',
-    'gameState',
-    'learners',
-    'meta',
-    'monsterVisualConfig',
-    'practiceSessions',
-    'revision',
-    'subjectStates',
-    'syncState',
-  ]),
-  requiredRevisionKeys: Object.freeze([
-    'accountRevision',
-    'accountLearnerListRevision',
-    'bootstrapCapacityVersion',
-    'hash',
-    'selectedLearnerRevision',
-  ]),
-});
+// PUBLIC_BOOTSTRAP_* constants, BOOTSTRAP_CAPACITY_VERSION,
+// BOOTSTRAP_MODES, BOOTSTRAP_V2_ENVELOPE_SHAPE → bootstrap-repository.js
 
 // U7: the classroom summary paginates at 50 learners per page (plan R11).
 const CLASSROOM_LEARNERS_SUMMARY_PAGE_LIMIT = 50;
@@ -205,54 +206,13 @@ const PUBLIC_EVENT_TEXT_ENUMS = {
   fromPhase: new Set(['retry', 'correction']),
 };
 
-function safeJsonParse(text, fallback) {
-  if (text == null || text === '') return cloneSerialisable(fallback);
-  try {
-    return JSON.parse(text);
-  } catch {
-    return cloneSerialisable(fallback);
-  }
-}
-
-function asTs(value, fallback) {
-  const numeric = Number(value);
-  if (Number.isFinite(numeric)) return numeric;
-  return fallback;
-}
-
-function isMissingTableError(error, tableName) {
-  const message = String(error?.message || '');
-  return new RegExp(`no such table:\\s*${tableName}\\b`, 'i').test(message);
-}
+// safeJsonParse, asTs, isMissingTableError → repository-helpers.js
 
 function isMissingCapacityReadModelTableError(error) {
   return CAPACITY_READ_MODEL_TABLES.some((tableName) => isMissingTableError(error, tableName));
 }
 
-function isPlainObject(value) {
-  return Boolean(value) && typeof value === 'object' && !Array.isArray(value);
-}
-
-function stableClone(value) {
-  if (Array.isArray(value)) return value.map(stableClone);
-  if (isPlainObject(value)) {
-    return Object.keys(value)
-      .sort()
-      .reduce((output, key) => {
-        output[key] = stableClone(value[key]);
-        return output;
-      }, {});
-  }
-  return value;
-}
-
-function stableStringify(value) {
-  return JSON.stringify(stableClone(cloneSerialisable(value)));
-}
-
-function mutationPayloadHash(kind, payload) {
-  return stableStringify({ kind, payload: cloneSerialisable(payload) });
-}
+// isPlainObject, stableClone, stableStringify, mutationPayloadHash → repository-helpers.js
 
 function subjectStateRowToRecord(row) {
   return normaliseSubjectStateRecord({
@@ -409,17 +369,7 @@ async function publicSubjectStateRowToRecord(row, { spellingContentSnapshot = nu
   });
 }
 
-function learnerRowToRecord(row) {
-  return normaliseLearnerRecord({
-    id: row.id,
-    name: row.name,
-    yearGroup: row.year_group,
-    avatarColor: row.avatar_color,
-    goal: row.goal,
-    dailyMinutes: row.daily_minutes,
-    createdAt: row.created_at,
-  }, row.id);
-}
+// learnerRowToRecord → membership-repository.js
 
 function gameStateRowToRecord(row) {
   return cloneSerialisable(safeJsonParse(row.state_json, {})) || {};
@@ -718,156 +668,11 @@ async function readSpellingRuntimeContentBundle(db, accountId, subjectId = 'spel
     || rememberSpellingRuntimeContent(key, buildSpellingRuntimeContent(row, subjectId));
 }
 
-function writableRole(role) {
-  return WRITABLE_MEMBERSHIP_ROLES.has(role);
-}
+// writableRole → membership-repository.js
 
-function normaliseMutationInput(rawValue, scopeType) {
-  const raw = isPlainObject(rawValue) ? rawValue : {};
-  const requestId = typeof raw.requestId === 'string' && raw.requestId ? raw.requestId : null;
-  const correlationId = typeof raw.correlationId === 'string' && raw.correlationId
-    ? raw.correlationId
-    : requestId;
-  const expectedRevisionKey = scopeType === 'account'
-    ? 'expectedAccountRevision'
-    : 'expectedLearnerRevision';
-  const expectedRevision = Number.isFinite(Number(raw[expectedRevisionKey]))
-    ? Number(raw[expectedRevisionKey])
-    : null;
-
-  if (!requestId) {
-    throw new BadRequestError('Mutation requestId is required for write routes.', {
-      code: 'mutation_request_id_required',
-      scopeType,
-    });
-  }
-
-  if (expectedRevision == null) {
-    throw new BadRequestError(`Mutation ${expectedRevisionKey} is required for write routes.`, {
-      code: 'mutation_revision_required',
-      scopeType,
-    });
-  }
-
-  return {
-    requestId,
-    correlationId,
-    expectedRevision,
-    expectedRevisionKey,
-  };
-}
-
-function buildMutationMeta({
-  kind,
-  scopeType,
-  scopeId,
-  requestId,
-  correlationId,
-  expectedRevision,
-  appliedRevision,
-  replayed = false,
-} = {}) {
-  return {
-    policyVersion: MUTATION_POLICY_VERSION,
-    kind,
-    scopeType,
-    scopeId,
-    requestId,
-    correlationId,
-    expectedRevision,
-    appliedRevision,
-    replayed,
-  };
-}
-
-function staleWriteError({ kind, scopeType, scopeId, requestId, correlationId, expectedRevision, currentRevision }) {
-  return new ConflictError('Mutation rejected because this state changed in another tab or device. Retry sync to reload the latest state, then repeat the action.', {
-    code: 'stale_write',
-    retryable: false,
-    kind,
-    scopeType,
-    scopeId,
-    requestId,
-    correlationId,
-    expectedRevision,
-    currentRevision,
-  });
-}
-
-function idempotencyReuseError({ kind, scopeType, scopeId, requestId, correlationId }) {
-  return new ConflictError('The same mutation request id was reused for a different payload.', {
-    code: 'idempotency_reuse',
-    retryable: false,
-    kind,
-    scopeType,
-    scopeId,
-    requestId,
-    correlationId,
-  });
-}
-
-function logMutation(level, event, details = {}) {
-  const payload = {
-    event,
-    ...cloneSerialisable(details),
-    at: new Date().toISOString(),
-  };
-  const fn = globalThis.console?.[level] || globalThis.console?.log;
-  if (!fn) return;
-  try {
-    fn('[ks2-worker]', JSON.stringify(payload));
-  } catch {
-    fn('[ks2-worker]', payload);
-  }
-}
-
-async function loadMutationReceipt(db, accountId, requestId) {
-  return first(db, `
-    SELECT account_id, request_id, scope_type, scope_id, mutation_kind, request_hash, response_json, status_code, correlation_id, applied_at
-    FROM mutation_receipts
-    WHERE account_id = ? AND request_id = ?
-  `, [accountId, requestId]);
-}
-
-async function storeMutationReceipt(db, {
-  accountId,
-  requestId,
-  scopeType,
-  scopeId,
-  mutationKind,
-  requestHash,
-  response,
-  statusCode = 200,
-  correlationId = null,
-  appliedAt,
-}) {
-  await run(db, `
-    INSERT INTO mutation_receipts (
-      account_id,
-      request_id,
-      scope_type,
-      scope_id,
-      mutation_kind,
-      request_hash,
-      response_json,
-      status_code,
-      correlation_id,
-      applied_at
-    )
-    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-  `, [
-    accountId,
-    requestId,
-    scopeType,
-    scopeId,
-    mutationKind,
-    requestHash,
-    JSON.stringify(response),
-    statusCode,
-    correlationId,
-    appliedAt,
-  ]);
-}
+// normaliseMutationInput, buildMutationMeta, staleWriteError,
+// idempotencyReuseError, loadMutationReceipt, storeMutationReceipt
+// → mutation-repository.js
 
 function storeMutationReceiptStatement(db, {
   accountId,
@@ -926,163 +731,13 @@ async function ensureAccount(db, session, nowTs) {
   `, [session.accountId, session.email, session.displayName, platformRole, nowTs, nowTs]);
 }
 
-async function listMembershipRows(db, accountId, { writableOnly = false } = {}) {
-  const allowedRoles = writableOnly ? ['owner', 'member'] : ['owner', 'member', 'viewer'];
-  const rolePlaceholders = sqlPlaceholders(allowedRoles.length);
-  return all(db, `
-    SELECT
-      m.account_id,
-      m.learner_id,
-      m.role,
-      m.sort_index,
-      m.created_at AS membership_created_at,
-      m.updated_at AS membership_updated_at,
-      l.id,
-      l.name,
-      l.year_group,
-      l.avatar_color,
-      l.goal,
-      l.daily_minutes,
-      l.created_at,
-      l.updated_at,
-      l.state_revision
-    FROM account_learner_memberships m
-    JOIN learner_profiles l ON l.id = m.learner_id
-    WHERE m.account_id = ?
-      AND m.role IN (${rolePlaceholders})
-    ORDER BY m.sort_index ASC, l.created_at ASC, l.id ASC
-  `, [accountId, ...allowedRoles]);
-}
-
-async function getMembership(db, accountId, learnerId) {
-  return first(db, `
-    SELECT account_id, learner_id, role, sort_index, created_at, updated_at
-    FROM account_learner_memberships
-    WHERE account_id = ? AND learner_id = ?
-  `, [accountId, learnerId]);
-}
-
-async function requireLearnerWriteAccess(db, accountId, learnerId) {
-  const membership = await getMembership(db, accountId, learnerId);
-  if (!membership || !writableRole(membership.role)) {
-    throw new ForbiddenError('Learner access denied.', {
-      learnerId,
-      required: 'owner-or-member',
-    });
-  }
-  return membership;
-}
-
-async function requireLearnerReadAccess(db, accountId, learnerId) {
-  const membership = await getMembership(db, accountId, learnerId);
-  if (!membership || !MEMBERSHIP_ROLES.has(membership.role)) {
-    throw new ForbiddenError('Learner access denied.', {
-      learnerId,
-      required: 'owner-member-or-viewer',
-    });
-  }
-  return membership;
-}
-
-function membershipRowToModel(row) {
-  return {
-    learnerId: row?.learner_id || row?.id || '',
-    role: row?.role || 'viewer',
-    sortIndex: Number(row?.sort_index) || 0,
-    stateRevision: Number(row?.state_revision) || 0,
-    learner: learnerRowToRecord(row),
-  };
-}
-
-function accountPlatformRole(account) {
-  return normalisePlatformRole(account?.platform_role);
-}
-
-function accountType(account) {
-  return account?.account_type === 'demo' ? 'demo' : 'real';
-}
-
-function requireParentHubAccess(account, membership) {
-  if (!canViewParentHub({ platformRole: accountPlatformRole(account), membershipRole: membership?.role })) {
-    throw new ForbiddenError('Parent Hub access denied.', {
-      code: 'parent_hub_forbidden',
-      required: 'platform-role-parent-or-admin plus readable learner membership',
-      learnerId: membership?.learner_id || null,
-    });
-  }
-}
-
-function requireAdminHubAccess(account) {
-  if (accountType(account) === 'demo' || !canViewAdminHub({ platformRole: accountPlatformRole(account) })) {
-    throw new ForbiddenError('Admin / operations access denied.', {
-      code: 'admin_hub_forbidden',
-      required: 'platform-role-admin-or-ops',
-    });
-  }
-}
-
-function requireAccountRoleManager(account) {
-  if (accountType(account) === 'demo' || !canManageAccountRoles({ platformRole: accountPlatformRole(account) })) {
-    throw new ForbiddenError('Account role management requires an admin account.', {
-      code: 'account_roles_forbidden',
-      required: 'platform-role-admin',
-    });
-  }
-}
-
-function requireMonsterVisualConfigManager(account) {
-  if (accountType(account) === 'demo' || accountPlatformRole(account) !== 'admin') {
-    throw new ForbiddenError('Monster visual config changes require an admin account.', {
-      code: 'monster_visual_config_forbidden',
-      required: 'platform-role-admin',
-    });
-  }
-}
-
-// U10 follower (MEDIUM — admin-only policy lock): the Grammar Writing
-// Try archive + hard-delete routes are destructive data mutations. The
-// reviewer convergence chose the stricter gate (admin only, ops 403)
-// rather than `requireAdminHubAccess` which grants ops through. Mirrors
-// `requireMonsterVisualConfigManager` and emits a dedicated error code
-// (`grammar_transfer_admin_forbidden`) so the security test can lock
-// the exact policy string.
-function requireGrammarTransferAdmin(account) {
-  if (accountType(account) === 'demo' || accountPlatformRole(account) !== 'admin') {
-    throw new ForbiddenError('Grammar Writing Try archive and delete require an admin account.', {
-      code: 'grammar_transfer_admin_forbidden',
-      required: 'platform-role-admin',
-    });
-  }
-}
-
-function requireSubjectContentExportAccess(account) {
-  if (!canViewAdminHub({ platformRole: accountPlatformRole(account) })) {
-    throw new ForbiddenError('Spelling content export requires an admin or operations account.', {
-      code: 'subject_content_export_forbidden',
-      required: 'platform-role-admin-or-ops',
-    });
-  }
-}
-
-function requireSubjectContentWriteAccess(account) {
-  if (accountPlatformRole(account) !== 'admin') {
-    throw new ForbiddenError('Spelling content import requires an admin account.', {
-      code: 'subject_content_write_forbidden',
-      required: 'platform-role-admin',
-    });
-  }
-}
-
-function normaliseRequestedPlatformRole(value) {
-  const role = typeof value === 'string' ? value.trim().toLowerCase() : '';
-  if (!PLATFORM_ROLES.includes(role)) {
-    throw new BadRequestError('Unknown platform role.', {
-      code: 'unknown_platform_role',
-      allowed: PLATFORM_ROLES,
-    });
-  }
-  return role;
-}
+// listMembershipRows, getMembership, requireLearnerWriteAccess,
+// requireLearnerReadAccess, membershipRowToModel, accountPlatformRole,
+// accountType, requireParentHubAccess, requireAdminHubAccess,
+// requireAccountRoleManager, requireMonsterVisualConfigManager,
+// requireGrammarTransferAdmin, requireSubjectContentExportAccess,
+// requireSubjectContentWriteAccess, normaliseRequestedPlatformRole
+// → membership-repository.js
 
 function runtimeSnapshotForBundle(bundle) {
   const backfilled = backfillSpellingWordExplanations(bundle, SEEDED_SPELLING_CONTENT_BUNDLE);
@@ -6377,82 +6032,9 @@ async function listPublicBootstrapEventRows(db, learnerIds) {
 //   bootstrapCapacityVersion:<V>;accountLearnerListRevision:<L>;
 //   writableLearnerStatesDigest:<D>
 //
-// The `accountId` prefix (U7 adv-u7-r1-002) salts the hash per account so
-// two accounts with identical (N,M,V,L,D) tuples no longer collide. The
-// `writableLearnerStatesDigest` slot (U1 follow-up 2026-04-26, B1 blocker)
-// pins EVERY writable learner's `learner_profiles.state_revision` into
-// the hash input, so a sibling (non-selected) `writeSubjectState` →
-// `withLearnerMutation` bump forces `bootstrapNotModifiedProbe` to miss
-// and the client gets a fresh full bundle. Without this slot, the
-// 4-input hash (N,M,V,L) was insensitive to sibling writes and the U1
-// hotfix was silently defeated when Nelson/James finished a round on a
-// second device while Eugenia was selected.
-//
-// Changing this input format (or the truncation length) is equivalent to
-// bumping `BOOTSTRAP_CAPACITY_VERSION` — stale clients will silently
-// reject `notModified` responses via the schema check. The version bump
-// in this PR from 2→3 forces v2 clients to miss once and re-bind.
-export async function computeBootstrapRevisionHash({
-  accountId,
-  accountRevision,
-  selectedLearnerRevision,
-  bootstrapCapacityVersion,
-  accountLearnerListRevision,
-  // U1 follow-up 2026-04-26: digest over every writable learner's
-  // `state_revision` (sorted by id, joined as `id:rev,id:rev`, SHA-256 →
-  // 16 bytes hex). Computed by `computeWritableLearnerStatesDigest()`.
-  // Optional for historical callers; defaults to `0` which matches the
-  // old 4-input hash when no digest is passed (migration-safe for any
-  // internal test that only exercises the raw hash helper).
-  writableLearnerStatesDigest = '',
-}) {
-  const input = [
-    `accountId:${String(accountId || '')}`,
-    `accountRevision:${Number(accountRevision) || 0}`,
-    `selectedLearnerRevision:${Number(selectedLearnerRevision) || 0}`,
-    `bootstrapCapacityVersion:${Number(bootstrapCapacityVersion) || 0}`,
-    `accountLearnerListRevision:${Number(accountLearnerListRevision) || 0}`,
-    `writableLearnerStatesDigest:${String(writableLearnerStatesDigest || '')}`,
-  ].join(';');
-  const bytes = new TextEncoder().encode(input);
-  const digest = await globalThis.crypto.subtle.digest('SHA-256', bytes);
-  const out = new Uint8Array(digest).slice(0, 16);
-  let hex = '';
-  for (let i = 0; i < out.length; i += 1) {
-    hex += out[i].toString(16).padStart(2, '0');
-  }
-  return hex;
-}
+// computeBootstrapRevisionHash → bootstrap-repository.js
 
-// U1 follow-up 2026-04-26 (B1 blocker): deterministic digest over every
-// writable learner's `state_revision`. Input rows are sorted by learner
-// id so addition/removal/reorder is captured by the upstream
-// `accountLearnerListRevision` rather than here. Per-learner bumps
-// (sibling `writeSubjectState` etc.) flow through this slot so the
-// `bootstrapNotModifiedProbe` short-circuit invalidates correctly.
-//
-// Returns 16 bytes hex (32 chars). Empty input returns an empty string;
-// `computeBootstrapRevisionHash` stamps that case as
-// `writableLearnerStatesDigest:` which is distinguishable from a real
-// digest.
-async function computeWritableLearnerStatesDigest(membershipRows) {
-  if (!Array.isArray(membershipRows) || !membershipRows.length) return '';
-  const entries = membershipRows
-    .map((row) => ({ id: String(row.id || ''), revision: Number(row.state_revision) || 0 }))
-    .filter((entry) => entry.id)
-    .sort((a, b) => (a.id < b.id ? -1 : a.id > b.id ? 1 : 0))
-    .map((entry) => `${entry.id}:${entry.revision}`)
-    .join(',');
-  if (!entries) return '';
-  const bytes = new TextEncoder().encode(entries);
-  const digest = await globalThis.crypto.subtle.digest('SHA-256', bytes);
-  const out = new Uint8Array(digest).slice(0, 16);
-  let hex = '';
-  for (let i = 0; i < out.length; i += 1) {
-    hex += out[i].toString(16).padStart(2, '0');
-  }
-  return hex;
-}
+// computeWritableLearnerStatesDigest → bootstrap-repository.js
 
 // U7: read the sibling-table counter. Missing row is revision 0 (lazy
 // creation deferred to the first bump). Tolerant of the helper table not
@@ -6490,111 +6072,11 @@ async function bumpAccountLearnerListRevision(db, accountId, nowTs) {
   }
 }
 
-// U7: compact `account.learnerList` entry for unselected learners in the
-// selected-learner-bounded response. Hard limit on per-entry payload —
-// no avatar blobs, no history, no prompts. Roughly 150 bytes per entry
-// after JSON serialisation; 50 entries → ~7.5 KB.
-function compactLearnerListEntry(row) {
-  if (!row) return null;
-  return {
-    id: String(row.id || ''),
-    name: String(row.name || ''),
-    avatarColor: row.avatar_color ? String(row.avatar_color) : null,
-    yearGroup: row.year_group ? String(row.year_group) : null,
-    revision: Number(row.state_revision) || 0,
-  };
-}
+// compactLearnerListEntry → bootstrap-repository.js
 
-function bootstrapCapacityMeta({
-  publicReadModels,
-  learnerCount,
-  sessionRows,
-  eventRows,
-  // U1 follow-up 2026-04-26 (B4): derive rather than hardcode. Caller
-  // must pass the actual query-shape boolean
-  // (`subjectStateLearnerIds.length === learnerIds.length` in the
-  // happy path; `false` when unbounded, `true` if a future re-bound
-  // mode is introduced or the B2 fallback shrinks the query). Keeping
-  // the parameter explicit means a future author cannot silently
-  // re-bound subject states without tripping both the caller shape
-  // and this contract.
-  subjectStatesBounded,
-  // U1 follow-up 2026-04-26 (B2): optional diagnostic stamped when the
-  // wider IN-clause degraded to the bounded fallback. `null` means the
-  // nominal path (every writable learner). `'degraded-to-selected'`
-  // means the widened SELECT failed and we fell back to [selectedId].
-  subjectStatesFallbackMode = null,
-}) {
-  if (!publicReadModels) return null;
-  const sessionActiveLimit = learnerCount * PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LIMIT_PER_LEARNER;
-  const sessionRecentLimit = learnerCount * PUBLIC_BOOTSTRAP_RECENT_SESSION_LIMIT_PER_LEARNER;
-  const sessionLimit = sessionActiveLimit + sessionRecentLimit;
-  const eventRecentLimit = learnerCount * PUBLIC_BOOTSTRAP_RECENT_EVENT_LIMIT_PER_LEARNER;
-  return {
-    version: PUBLIC_BOOTSTRAP_CAPACITY_VERSION,
-    mode: 'public-bounded',
-    limits: {
-      activeSessionsPerLearner: PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LIMIT_PER_LEARNER,
-      recentSessionsPerLearner: PUBLIC_BOOTSTRAP_RECENT_SESSION_LIMIT_PER_LEARNER,
-      recentEventsPerLearner: PUBLIC_BOOTSTRAP_RECENT_EVENT_LIMIT_PER_LEARNER,
-    },
-    learners: {
-      returned: learnerCount,
-    },
-    practiceSessions: {
-      returned: sessionRows.length,
-      bounded: true,
-      atOrAboveRecentLimit: learnerCount > 0 && sessionRows.length >= sessionRecentLimit,
-      atOrAboveMaximumLimit: learnerCount > 0 && sessionRows.length >= sessionLimit,
-    },
-    eventLog: {
-      returned: eventRows.length,
-      bounded: true,
-      atOrAboveRecentLimit: learnerCount > 0 && eventRows.length >= eventRecentLimit,
-    },
-    // U1 hotfix 2026-04-26: child_subject_state + child_game_state ship for
-    // every writable learner even in selected-learner-bounded mode, so the
-    // Spelling/Grammar/Punctuation "Where You Stand" setup stats no longer
-    // show 0 for non-selected learners. Spec:
-    // docs/superpowers/specs/2026-04-26-bootstrap-learner-stats-hotfix-
-    // design.md. U1 follow-up (B4): value is derived from the caller's
-    // actual query shape — a drift-prone hardcoded `false` would silently
-    // lie if a future author re-bounded subject states.
-    subjectStatesBounded: Boolean(subjectStatesBounded),
-    // U1 follow-up (B2) defensive fallback. Omitted (undefined) when the
-    // widened SELECTs succeeded; `'degraded-to-selected'` when they
-    // tripped a D1 failure and we fell back to [selectedId] so the
-    // bootstrap still returns rather than 500s.
-    ...(subjectStatesFallbackMode ? { subjectStatesFallbackMode } : {}),
-  };
-}
+// bootstrapCapacityMeta → bootstrap-repository.js
 
-// U7: resolve the "cold-start" selected learner given optional client
-// preference. Precedence (per plan line 756):
-//   1. preferredLearnerId (if writable in caller's scope)
-//   2. persisted account.selected_learner_id (if still writable)
-//   3. first alphabetical by learner id
-// Client preference pointing at a non-writable id is silently rejected —
-// do NOT leak `clientPreferenceRejected` in the response body per plan
-// line 778.
-function resolveBootstrapSelectedLearnerId(
-  membershipRows,
-  persistedSelectedId,
-  preferredLearnerId,
-) {
-  const writableIds = new Set(
-    membershipRows.filter((row) => writableRole(row.role)).map((row) => String(row.id)),
-  );
-  if (!writableIds.size) return null;
-  const preferred = preferredLearnerId ? String(preferredLearnerId) : '';
-  if (preferred && writableIds.has(preferred)) return preferred;
-  if (persistedSelectedId && writableIds.has(String(persistedSelectedId))) {
-    return String(persistedSelectedId);
-  }
-  // Alphabetical fallback.
-  const sorted = [...writableIds].sort();
-  return sorted[0] || null;
-}
+// resolveBootstrapSelectedLearnerId → bootstrap-repository.js
 
 async function bootstrapBundle(db, accountId, {
   publicReadModels = false,
@@ -7557,236 +7039,7 @@ async function writeLearnersSnapshot(db, accountId, snapshot, nowTs) {
   return bootstrapBundle(db, accountId);
 }
 
-async function withAccountMutation(db, {
-  accountId,
-  kind,
-  payload,
-  mutation,
-  nowTs,
-  apply,
-  receiptResponse = (response) => response,
-  replayResponse = null,
-}) {
-  const nextMutation = normaliseMutationInput(mutation, 'account');
-  const requestHash = mutationPayloadHash(kind, payload);
-
-  // NOTE: non-atomic by design — (a) branching on intermediate read results
-  // (existingReceipt short-circuit, repo_revision CAS compare) plus (b) an
-  // `apply()` callback that runs its own write path. `withTransaction` was
-  // removed in U12 (production D1 no-op). The CAS UPDATE itself
-  // (`WHERE repo_revision = ?`) is the authoritative stale-write defence.
-  return (async () => {
-    const existingReceipt = await loadMutationReceipt(db, accountId, nextMutation.requestId);
-    if (existingReceipt) {
-      if (existingReceipt.request_hash !== requestHash) {
-        throw idempotencyReuseError({
-          kind,
-          scopeType: 'account',
-          scopeId: accountId,
-          requestId: nextMutation.requestId,
-          correlationId: nextMutation.correlationId,
-        });
-      }
-      const storedReplay = safeJsonParse(existingReceipt.response_json, {});
-      const replayed = typeof replayResponse === 'function'
-        ? await replayResponse({ storedReplay, existingReceipt, mutation: nextMutation })
-        : storedReplay;
-      replayed.mutation = buildMutationMeta({
-        ...replayed.mutation,
-        kind,
-        scopeType: 'account',
-        scopeId: accountId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        replayed: true,
-      });
-      logMutation('info', 'mutation.replayed', {
-        kind,
-        scopeType: 'account',
-        scopeId: accountId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-      });
-      return replayed;
-    }
-
-    const account = await first(db, 'SELECT id, repo_revision FROM adult_accounts WHERE id = ?', [accountId]);
-    if (!account) throw new NotFoundError('Account scope was not found.', { accountId });
-
-    const casMeta = await run(db, `
-      UPDATE adult_accounts
-      SET repo_revision = repo_revision + 1,
-          updated_at = ?
-      WHERE id = ?
-        AND repo_revision = ?
-    `, [nowTs, accountId, nextMutation.expectedRevision]);
-    const casChanges = Number(casMeta?.meta?.changes) || 0;
-    if (casChanges !== 1) {
-      const currentRevision = Number(await scalar(db, 'SELECT repo_revision FROM adult_accounts WHERE id = ?', [accountId], 'repo_revision')) || 0;
-      throw staleWriteError({
-        kind,
-        scopeType: 'account',
-        scopeId: accountId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        expectedRevision: nextMutation.expectedRevision,
-        currentRevision,
-      });
-    }
-
-    const appliedRevision = nextMutation.expectedRevision + 1;
-    const applied = await apply();
-    const response = {
-      ...applied,
-      mutation: buildMutationMeta({
-        kind,
-        scopeType: 'account',
-        scopeId: accountId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        expectedRevision: nextMutation.expectedRevision,
-        appliedRevision,
-      }),
-    };
-    await storeMutationReceipt(db, {
-      accountId,
-      requestId: nextMutation.requestId,
-      scopeType: 'account',
-      scopeId: accountId,
-      mutationKind: kind,
-      requestHash,
-      response: receiptResponse(response),
-      correlationId: nextMutation.correlationId,
-      appliedAt: nowTs,
-    });
-    logMutation('info', 'mutation.applied', {
-      kind,
-      scopeType: 'account',
-      scopeId: accountId,
-      requestId: nextMutation.requestId,
-      correlationId: nextMutation.correlationId,
-      expectedRevision: nextMutation.expectedRevision,
-      appliedRevision,
-    });
-    return response;
-  })();
-}
-
-async function withLearnerMutation(db, {
-  accountId,
-  learnerId,
-  kind,
-  payload,
-  mutation,
-  nowTs,
-  apply,
-}) {
-  if (!(typeof learnerId === 'string' && learnerId)) {
-    throw new BadRequestError('Learner id is required for this mutation.', { code: 'learner_id_required', kind });
-  }
-
-  const nextMutation = normaliseMutationInput(mutation, 'learner');
-  const requestHash = mutationPayloadHash(kind, payload);
-
-  // NOTE: non-atomic by design — (a) branching on intermediate read results
-  // (write-access check, existingReceipt short-circuit, state_revision CAS
-  // compare) plus (b) an `apply()` callback that runs its own write path.
-  // `withTransaction` was removed in U12 (silent production no-op). The
-  // CAS UPDATE (`WHERE state_revision = ?`) is the stale-write defence.
-  return (async () => {
-    await requireLearnerWriteAccess(db, accountId, learnerId);
-    const existingReceipt = await loadMutationReceipt(db, accountId, nextMutation.requestId);
-    if (existingReceipt) {
-      if (existingReceipt.request_hash !== requestHash) {
-        throw idempotencyReuseError({
-          kind,
-          scopeType: 'learner',
-          scopeId: learnerId,
-          requestId: nextMutation.requestId,
-          correlationId: nextMutation.correlationId,
-        });
-      }
-      const replayed = safeJsonParse(existingReceipt.response_json, {});
-      replayed.mutation = buildMutationMeta({
-        ...replayed.mutation,
-        kind,
-        scopeType: 'learner',
-        scopeId: learnerId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        replayed: true,
-      });
-      logMutation('info', 'mutation.replayed', {
-        kind,
-        scopeType: 'learner',
-        scopeId: learnerId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-      });
-      return replayed;
-    }
-
-    const learner = await first(db, 'SELECT id FROM learner_profiles WHERE id = ?', [learnerId]);
-    if (!learner) throw new NotFoundError('Learner was not found.', { learnerId });
-
-    const casMeta = await run(db, `
-      UPDATE learner_profiles
-      SET state_revision = state_revision + 1,
-          updated_at = ?
-      WHERE id = ?
-        AND state_revision = ?
-    `, [nowTs, learnerId, nextMutation.expectedRevision]);
-    const casChanges = Number(casMeta?.meta?.changes) || 0;
-    if (casChanges !== 1) {
-      const currentRevision = Number(await scalar(db, 'SELECT state_revision FROM learner_profiles WHERE id = ?', [learnerId], 'state_revision')) || 0;
-      throw staleWriteError({
-        kind,
-        scopeType: 'learner',
-        scopeId: learnerId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        expectedRevision: nextMutation.expectedRevision,
-        currentRevision,
-      });
-    }
-
-    const appliedRevision = nextMutation.expectedRevision + 1;
-    const applied = await apply();
-    const response = {
-      ...applied,
-      mutation: buildMutationMeta({
-        kind,
-        scopeType: 'learner',
-        scopeId: learnerId,
-        requestId: nextMutation.requestId,
-        correlationId: nextMutation.correlationId,
-        expectedRevision: nextMutation.expectedRevision,
-        appliedRevision,
-      }),
-    };
-    await storeMutationReceipt(db, {
-      accountId,
-      requestId: nextMutation.requestId,
-      scopeType: 'learner',
-      scopeId: learnerId,
-      mutationKind: kind,
-      requestHash,
-      response,
-      correlationId: nextMutation.correlationId,
-      appliedAt: nowTs,
-    });
-    logMutation('info', 'mutation.applied', {
-      kind,
-      scopeType: 'learner',
-      scopeId: learnerId,
-      requestId: nextMutation.requestId,
-      correlationId: nextMutation.correlationId,
-      expectedRevision: nextMutation.expectedRevision,
-      appliedRevision,
-    });
-    return response;
-  })();
-}
+// withAccountMutation, withLearnerMutation → mutation-repository.js
 
 async function runSubjectCommandMutation(db, {
   accountId,
@@ -9161,3 +8414,43 @@ export function createWorkerRepository({ env = {}, now = Date.now, capacity = nu
     },
   };
 }
+
+// ─── Barrel re-exports ───────────────────────────────────────────────────────
+// P3 U6: extracted modules re-exported so existing consumers continue to
+// import from './repository.js' without import-path changes.
+
+export {
+  accountPlatformRole,
+  accountType,
+  getMembership,
+  learnerRowToRecord,
+  listMembershipRows,
+  membershipRowToModel,
+  normaliseRequestedPlatformRole,
+  requireAccountRoleManager,
+  requireAdminHubAccess,
+  requireGrammarTransferAdmin,
+  requireLearnerReadAccess,
+  requireLearnerWriteAccess,
+  requireMonsterVisualConfigManager,
+  requireParentHubAccess,
+  requireSubjectContentExportAccess,
+  requireSubjectContentWriteAccess,
+  writableRole,
+} from './membership-repository.js';
+
+export {
+  bootstrapCapacityMeta,
+  BOOTSTRAP_CAPACITY_VERSION,
+  BOOTSTRAP_MODES,
+  BOOTSTRAP_V2_ENVELOPE_SHAPE,
+  compactLearnerListEntry,
+  computeBootstrapRevisionHash,
+  computeWritableLearnerStatesDigest,
+  PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_ACTIVE_SESSION_LOOKUP_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_CAPACITY_VERSION,
+  PUBLIC_BOOTSTRAP_RECENT_EVENT_LIMIT_PER_LEARNER,
+  PUBLIC_BOOTSTRAP_RECENT_SESSION_LIMIT_PER_LEARNER,
+  resolveBootstrapSelectedLearnerId,
+} from './bootstrap-repository.js';
